refactor(pipeline): extract review_test.go from engine_test.go [#288]
Move parallel-review tests (TestEngine_ParallelReview_*, TestComputeReviewVerdict,
TestEngine_LoadPriorReview, TestNeededReviewersFromPrior, TestPriorFindingsForReviewer)
into a dedicated review_test.go file alongside review.go.
